+++ src/main/java/svenhjol/charmony/core/client/features/control_panel/FeaturesList.java
@@ -58,7 +58,9 @@ public class FeaturesList extends AbstractSelectionList<FeaturesList.Entry> {
 
     protected List<Feature> features() {
         if (features.isEmpty()) {
-            features.addAll(mod.features());
+            features.addAll(mod.features().stream()
+                .filter(Feature::showInConfig)
+                .toList());
         }
         return features;
     }